feat(eval): move binary expression logic to separate file
